WIP: encoding `gas_used` and having a refactoring for encoding/decoding wasm values as a by-product.
