+++ crates/svm-app/src/lib.rs
@@ -1,7 +1,7 @@
-#![deny(missing_docs)]
-#![deny(unused)]
-#![deny(dead_code)]
-#![deny(unreachable_code)]
+#![allow(missing_docs)]
+#![allow(unused)]
+#![allow(dead_code)]
+#![allow(unreachable_code)]
 
 //! The `svm-app` crate is responsible on `Spacemesh` `AppTemplate`(s) and `App`(s).
