+++ crates/svm-app/src/raw/gas/decoder.rs
@@ -1,0 +1,5 @@
+use super::super::NibbleIter;
+
+pub fn decode_gas_used(iter: &mut NibbleIter) {
+    //
+}
