+++ crates/svm-app/src/raw/gas/encoder.rs
@@ -1,0 +1,5 @@
+use super::super::NibbleWriter;
+
+pub fn encode_gas_used(gas_used: u64, w: &mut NibbleWriter) {
+    //
+}
