+++ crates/svm-app/src/raw/gas/mod.rs
@@ -1,0 +1,4 @@
+mod decoder;
+mod encoder;
+
+pub use encoder::encode_gas_used;
