+++ crates/svm-runtime-c-api/src/receipt/deploy_template.rs
@@ -1,11 +1,15 @@
 //!     `Deploy Template` Receipt Raw Format Version 0
 //!
 //!  On success (`is_success = 1`)
-//!  ----------------------------------------------------
+//!  +---------------------------------------------------+
 //!  |   format   |              |                       |
 //!  |  version   |  is_success  |    Template Address   |
 //!  |  (4 bytes) |   (1 byte)   |      (20 bytes)       |
-//!  |____________|______________|_______________________|
+//!  +____________|______________|_______________________+
+//!  |                                                   |
+//!  |                   gas_used                        |
+//!  +___________________________________________________+
+//!
 //!
 //!  On success (`is_success = 0`)
 //!  See [error.rs][./error.rs]
